+++ src/components/ConsoleBox/consoleBox.css
@@ -8,8 +8,8 @@
     overflow-x: wrap;
     word-break:break-all;
     font-family: 'Courier New', Courier, monospace;
-    transition-property: all;
-    transition-duration: .2s;
+    transition: all .2s ease;
+    /* transition-duration: .2s; */
     transition-delay: .25s;
 }
 .resize{
