+++ src/components/DisplayBox/card-list.css
