+++ src/components/SettingsBox/SettingsBox.css
@@ -1,5 +1,5 @@
 
-.skadingle {
+.setting-input {
     margin: 1px 0 10px 0;
     border-color: #999;
     border-width: 1px;
